+++ R/cpma.R
@@ -5,33 +5,33 @@
 #' This function allows you to run the cross-phenotype meta-analysis (CPMA)
 #' method given a variety of parameters.
 #'
-#' @param p_in test
+#' @param p test
 #' @param epsilon test
 #' @param .THRESHOLD test
 #' @param .MAXVAL test
 #' @param .FIXED test
 #' @export
 
-cpma <- function(p_in, epsilon = 0.001, .THRESHOLD = 5E-8, .MAXVAL = 1, .FIXED = TRUE) {
+cpma <- function(p, epsilon = 0.001, .THRESHOLD = 5E-8, .MAXVAL = 1, .FIXED = TRUE) {
 
-  # prep p_in
-  if (!is.list(p_in)) stop("`p_in` needs to be a list.")
-  m <- length(p_in)
-  if (m < 2) stop("`p_in` should have at least 2 elements (i.e., multiple p-value vectors).")
-  l <- sapply(p_in, length)
-  if (any(l != l[1])) stop("Not all elements of `p_in` are equally long.")
-  p <- do.call("cbind", p_in)
-  if (any(p < 0)) stop("Negative p-values detected, which is not possible.")
-  p[p > .MAXVAL] <- 1
+  # check input arguments
+  p_args <- check_p(p, .MAXVAL)
+  m <- p_args$number_of_traits
+  l <- p_args$number_of_snps
+  pm <- p_args$p_matrix
+
+  if (!is.numeric(epsilon) || epsilon < 0 || epsilon > 1)
+    stop("Make sure `epsilon` is a number between 0 and 1")
+
+  if (!is.numeric(.THRESHOLD) || .THRESHOLD < 0 || .THRESHOLD > 1)
+    stop("Make sure `.THRESHOLD` is a number between 0 and 1")
+
+  if (!is.logical(.FIXED)) stop("`.FIXED` should be `TRUE` or `FALSE`")
 
   # run
-  loi <- 0
-  final_p <- rep(NA, l[1])
-  final_n <- rep(NA, l[1])
-  final_traits <- vector(mode = "list", length = l[1])
-
+  final <- make_final(l)
   for (i in seq_len(l[1])) {
-    x <- p[i, ]
+    x <- pm[i, ]
     o <- fastorder(x)
     ms <- 1 / (cumsum(-log(o$x)) / 1:m)
     an <- -log(o$x) - epsilon
@@ -52,11 +52,10 @@ cpma <- function(p_in, epsilon = 0.001, .THRESHOLD = 5E-8, .MAXVAL = 1, .FIXED =
       traits <- o$ix[1:nn]
     }
 
-    final_p[i] <- p_out
-    final_n[i] <- n2
-    final_traits[[i]] <- traits
+    final$p[i] <- min(p_out, 1)
+    final$n[i] <- n2
+    final$traits[[i]] <- traits
+    final$p_exp[i] <- -log10(final$p[i])
   }
-
-  final <- list(p = pmin(final_p, 1), n = final_n, traits = final_traits)
   return(final)
 }