+++ tests/testthat/test-fct.R
@@ -1,15 +1,23 @@
 # default behavior
 p_2 <- list(0.05, 0.05)
 p_10 <- as.list(rep(0.05, 10))
+p_ns <- as.list(rep(5E-8, 10))
 
 test_that("fct works for two p-values", {
   expect_equal(fct(p_2)$p, 0.01747866, tolerance = 1E-5)
   expect_equal(fct(p_2)$n, 0)
+  expect_equal(fct(p_2)$p_exp, -log10(0.01747866), tolerance = 1E-5)
 })
 
 test_that("fct works for ten p-values", {
   expect_equal(fct(p_10)$p, 7.341634e-06, tolerance = 1E-5)
   expect_equal(fct(p_10)$n, 0)
+  expect_equal(fct(p_10)$p_exp, -log10(7.341634e-06), tolerance = 1E-5)
+})
+
+test_that("fct works the same as non-subset fct", {
+  expect_equal(fct(p_2)$p, do.call("p_fct", p_2))
+  expect_equal(fct(p_ns)$p, do.call("p_fct", p_ns))
 })
 
 # handling exceptions
@@ -25,4 +33,23 @@ test_that("fct can deal with illogical values", {
   expect_error(fct(p_nan))
   expect_error(fct(p_neg))
   expect_error(fct(p_char))
+
+  expect_error(fct(p_2, .THRESHOLD = "test"))
+  expect_error(fct(p_2, .THRESHOLD = -1))
+  expect_error(fct(p_2, .THRESHOLD = 2))
+
+  expect_error(fct(p_2, .MAXVAL = "test"))
+  expect_error(fct(p_2, .MAXVAL = -1))
+  expect_error(fct(p_2, .MAXVAL = 2))
+})
+
+# handling extreme values
+p_0 <- list()
+p_1 <- list(0.05)
+p_10k <- as.list(rep(0.3, 10000))
+
+test_that("fct can deal with extreme values", {
+  expect_error(fct(p_0))
+  expect_error(fct(p_1))
+  expect_equal(fct(p_10k)$p, 4.113606e-82, tolerance = 1E-5)
 })